exynos-powermode: fix function signature
exynos_get_eint_wakemask returns an u32 but is declared as returning u64
here.

Caught by LTO:
include/soc/samsung/exynos-powermode.h:110: warning: type of 'exynos_get_eint_wake_mask' does not match original declaration [-Wlto-type-mismatch]
 extern u64 exynos_get_eint_wake_mask(void);

drivers/pinctrl/samsung/pinctrl-exynos.c:398: note: return value type mismatch
 u32 exynos_get_eint_wake_mask(void)

drivers/pinctrl/samsung/pinctrl-exynos.c:398: note: type 'u32' should match type 'u64'
drivers/pinctrl/samsung/pinctrl-exynos.c:398: note: 'exynos_get_eint_wake_mask' was previously declared here
drivers/pinctrl/samsung/pinctrl-exynos.c:398: note: code may be misoptimized unless -fno-strict-aliasing is used

+++ include/soc/samsung/exynos-powermode.h
@@ -107,9 +107,9 @@ static inline void s3c24xx_serial_fifo_wait(void) { }
 #endif
 
 #ifdef CONFIG_PINCTRL_EXYNOS
-extern u64 exynos_get_eint_wake_mask(void);
+extern u32 exynos_get_eint_wake_mask(void);
 #else
-static inline u64 exynos_get_eint_wake_mask(void) { return 0xffffffffL; }
+static inline u32 exynos_get_eint_wake_mask(void) { return 0xffffffffL; }
 #endif
 
 /* SUPPORT HOTPLUG */
